+++ internal/statistics/service.go
@@ -6,10 +6,6 @@ import (
 	"github.com/certimate-go/certimate/internal/domain"
 )
 
-type statisticsRepository interface {
-	Get(ctx context.Context) (*domain.Statistics, error)
-}
-
 type StatisticsService struct {
 	statRepo statisticsRepository
 }
