+++ internal/workflow/service_deps.go
@@ -1,0 +1,25 @@
+package workflow
+
+import (
+	"context"
+
+	"github.com/pocketbase/dbx"
+
+	"github.com/certimate-go/certimate/internal/domain"
+)
+
+type workflowRepository interface {
+	ListEnabledScheduled(ctx context.Context) ([]*domain.Workflow, error)
+	GetById(ctx context.Context, id string) (*domain.Workflow, error)
+	Save(ctx context.Context, workflow *domain.Workflow) (*domain.Workflow, error)
+}
+
+type workflowRunRepository interface {
+	GetById(ctx context.Context, id string) (*domain.WorkflowRun, error)
+	SaveWithCascading(ctx context.Context, workflowRun *domain.WorkflowRun) (*domain.WorkflowRun, error)
+	DeleteWhere(ctx context.Context, exprs ...dbx.Expression) (int, error)
+}
+
+type settingsRepository interface {
+	GetByName(ctx context.Context, name string) (*domain.Settings, error)
+}
